Collision is bugged will implement a new method of collision based on distances

+++ src/objects/player.cpp
@@ -1,6 +1,6 @@
 #include "objects.hpp"
 
-double Player::Player::CalculateFeet()
+long double Player::Player::CalculateFeet()
 {
     return (bodyPart.ReturnPositionY() - bodyPart.ReturnHeight() / 2);
 }
